# upravljanjevremenom/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Time Manager API",
        "description": "REST API sistem za upravljanje vremenom. \nOmogućava korisnicima upravljanje zadacima, beleškama i obaveštenjima, \npraćenje statusa zadataka, kao i pregled statistike rada korisnika. \nSistem podržava uloge (admin, regular i premium korisnik).",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/zadaci": {
            "get": {
                "tags": [
                    "Zadaci"
                ],
                "summary": "Lista svih zadataka",
                "operationId": "9d6f49b88a357c35f037d9f1d21d1b03",
                "responses": {
                    "200": {
                        "description": "Uspešno vraćena lista zadataka"
                    }
                }
            },
            "post": {
                "tags": [
                    "Zadaci"
                ],
                "summary": "Kreiranje novog zadatka",
                "operationId": "c050cc8c27718ee1a39622d74beb7f9e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "naslov",
                                    "korisnik_id"
                                ],
                                "properties": {
                                    "naslov": {
                                        "type": "string",
                                        "example": "Završiti projekat"
                                    },
                                    "opis": {
                                        "type": "string",
                                        "example": "Implementirati Swagger dokumentaciju"
                                    },
                                    "uradjeno": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "rok": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-02-25 23:59:00"
                                    },
                                    "korisnik_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Zadatak uspešno kreiran"
                    },
                    "422": {
                        "description": "Validaciona greška"
                    }
                }
            }
        },
        "/api/zadaci/{id}": {
            "get": {
                "tags": [
                    "Zadaci"
                ],
                "summary": "Prikaz jednog zadatka",
                "operationId": "1d0937b62673460777ffed4821b9de12",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Zadatak pronađen"
                    },
                    "404": {
                        "description": "Zadatak nije pronađen"
                    }
                }
            },
            "put": {
                "tags": [
                    "Zadaci"
                ],
                "summary": "Izmena zadatka",
                "operationId": "ae554b9cfa49ad2d13ff06125dcd544a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "naslov": {
                                        "type": "string",
                                        "example": "Izmenjen naslov"
                                    },
                                    "opis": {
                                        "type": "string",
                                        "example": "Izmenjen opis"
                                    },
                                    "uradjeno": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "rok": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-02-28 18:00:00"
                                    },
                                    "korisnik_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Zadatak uspešno izmenjen"
                    },
                    "404": {
                        "description": "Zadatak nije pronađen"
                    },
                    "422": {
                        "description": "Validaciona greška"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Zadaci"
                ],
                "summary": "Brisanje zadatka",
                "operationId": "0cf9b9a5586999f9ca3c2855de39aee9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Zadatak uspešno obrisan"
                    },
                    "404": {
                        "description": "Zadatak nije pronađen"
                    }
                }
            }
        }
    },
...
}